Add code to load FAT clusters from disk.
This commit adds a short subroutine to load a single FAT cluster from the disk. In the future, another subroutine will be added which searches the FAT for the next cluster, so traversing cluster chains will be possible. This also removes some printing subroutines to save space. They may be replaced if there is room once the bootloader is working.