Add FAT32 subroutines and improve comments.
This commit introduces several subroutines (most notably find_file) for working with a FAT32 file system, with the end goal of loading the second stage of the bootloader from the disk. The next_cluster subroutine is the only thing yet to be implemented, and once it is implemented stage 1 of the bootloader will be done, from a functionality standpoint. This commit also increases the number and quality of comments throughout the code, hopefully making it more comprehensible and maintainable.