Implement next_cluster subroutine.
This commit implements the next_cluster subroutine, which allows for finding the address of the next cluster in a cluster chain, therefore making it possible to load files/directories of arbitrary length from the file system. A small error is also corrected, where the bootloader attempted to load stage 2 to 0x7E00. In the event that the second stage is stored across multiple clusters, this will fail, as information needed to load clusters from the disk is stored in the region 0x7E00-0x8000, and an attempt to load another cluster would overwrite 0x8000-0x8200 with a sector of the FAT. As such, stage 2 is now loaded to 0x8200.