Enable A20 line and load basic GDT.
This commit introduces code which will attempt to enable the A20 line
using INT 15h in the event that it is not already enabled, as well as
code which will load a basic GDT once the A20 line is enabled. Further
attempts to enable the A20 line using the keyboard controller method and
the fast A20 method must be implemented in the event that INT 15h fails.
A TSS must also be added to the GDT.